+++ Editor/Visualize/ReferenceInspectorGUI.cs
@@ -13,9 +13,18 @@ namespace RV
 
 		private static void EditorOnfinishedDefaultHeaderGUI(Editor editor)
 		{
+			if (!ReferenceSetting.IsEnabled) return;
+			
 			if (editor.targets.Length == 1)
 			{
 				Object target = editor.target;
+				if (!ReferenceSetting.TraceSceneObject && target is GameObject go)
+				{
+					if (go.IsSceneObject())
+					{
+						return;
+					}
+				}
 
 				string path = AssetDatabase.GetAssetPath(target);
 				string guid = AssetDatabase.AssetPathToGUID(path);